Fix leak in gluon.Trainer
